Extract serve command execution
